third_party/blink/web_tests/fullscreen/full-screen-prefixed-and-unprefixed-expected.txt

EVENT(webkitfullscreenchange)
EXPECTED (document.fullscreenElement == '[object HTMLDivElement]') OK
EXPECTED (document.webkitFullscreenElement == '[object HTMLDivElement]') OK
EVENT(fullscreenchange)
EXPECTED (document.fullscreenElement == '[object HTMLSpanElement]') OK
EXPECTED (document.webkitFullscreenElement == '[object HTMLSpanElement]') OK
EVENT(fullscreenchange)
EXPECTED (document.fullscreenElement == '[object HTMLDivElement]') OK
EXPECTED (document.webkitFullscreenElement == '[object HTMLDivElement]') OK
EVENT(webkitfullscreenchange)
EXPECTED (document.fullscreenElement == 'null') OK
EXPECTED (document.webkitFullscreenElement == 'null') OK
END OF TEST
